@@ -1,14 +1,11 @@
-source "https://rubygems.org"
+source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-gem "rails", "3.2.22.4"
+ruby '3.0.2'
 
-gem "sqlite3"
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
+gem 'rails', '~> 6.1.4', '>= 6.1.4.1'
+# Use sqlite3 as the database for Active Record
+gem 'sqlite3', '~> 1.4'
 
 gem "suture", :path => "../.."
-
-gem "minitest", :group => :test
-
-# 1.8.7 support
-gem "rake", "~> 10.0"
-gem "i18n", "~> 0.6.0"
-gem "rack-cache", "~> 1.2.0"
